refactor(symengine): finish the seFromSE decomposition
Clears the remaining habit-hooks findings on the new code.  Seven units, one
sentence of responsibility each:

  seFromSEarena.h  hold every string the emitter builds
  seFromSEnode.h   what kind of parse node is this, and what text does it span
  seFromSEnames.h  the name and constant vocabulary shared with symengine
  seFromSEfold.h   would R's baseenv() eval have folded this operand
  seFromSEcalls.h  which function calls the emitter can render, and how
  seFromSEemit.h   walk a parse tree and emit rxode2 text
  seFromSE.c       drive the parser and hand a batch back to R

seFromSEnode.h is the one that earned its place: the grammar spells the
precedence ladder out, so seIsBareNumber(), seStripP() and seFold() each
carried their own copy of the single-child pass-through productions.

Also lifted seFoldTernary() and seCallArgs() out of the two functions that
were still branchy, and made inFile an explicit argument to the fixture
capture helpers instead of a reach into the generator's scope.

What habit-hooks still reports is not actionable: 30 of the 34 duplicated-code
hits are in src/seFromSE.g.d_parser.h, which dparser::mkdparse() generates and
CLAUDE.md forbids editing (its parse tables are repetitive by construction);
the other four are seFromSE.c's own 43-line entry point matched fuzzily
against them.

No behavior change: 434 of 592 fixture expressions accepted, zero byte
differences, and the generator still reproduces a byte-identical fixture.
